Refactored deribit backend into new multi file format
